add shutdown at heartbeat and extract kv client
